enhance: Use SearchResult in MilvusClient
SearchResult's type is SearchResult[Hits[hit]], which is List[List[Hit]].

This PR changes `Hit` into `Dict`. So the SearchResult type becomes
List[List[Dict]] to match MilvusClient's return type.

This avoids extra copy in the MilvusClient search to enhance performance.
The usage of MilvusClient.search and orm.Collection.search keeps unchange.
