+++ pymilvus/client/abstract.py
@@ -1,13 +1,10 @@
 import abc
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional, Union
 
-import ujson
-
-from pymilvus.exceptions import DataTypeNotMatchException, ExceptionsMessage, MilvusException
-from pymilvus.grpc_gen import common_pb2, schema_pb2
+from pymilvus.exceptions import DataTypeNotMatchException, ExceptionsMessage
 from pymilvus.settings import Config
 
-from . import entity_helper, utils
+from . import utils
 from .constants import DEFAULT_CONSISTENCY_LEVEL, RANKER_TYPE_RRF, RANKER_TYPE_WEIGHTED
 from .types import DataType, FunctionType
 
@@ -455,379 +452,6 @@ class AnnSearchRequest:
             "limit": self.limit,
             "expr": self.expr,
         }.__str__()
-
-
-class SearchResult(list):
-    """nq results: List[Hits]"""
-
-    def __init__(
-        self,
-        res: schema_pb2.SearchResultData,
-        round_decimal: Optional[int] = None,
-        status: Optional[common_pb2.Status] = None,
-        session_ts: Optional[int] = 0,
-    ):
-        self._nq = res.num_queries
-        all_topks = res.topks
-        self.recalls = res.recalls
-        self._pk_name = res.primary_field_name or "id"
-
-        self.cost = int(status.extra_info["report_value"] if status and status.extra_info else "0")
-
-        output_fields = res.output_fields
-        fields_data = res.fields_data
-
-        all_pks: List[Union[str, int]] = []
-        all_scores: List[float] = []
-
-        if res.ids.HasField("int_id"):
-            all_pks = res.ids.int_id.data
-        elif res.ids.HasField("str_id"):
-            all_pks = res.ids.str_id.data
-
-        if isinstance(round_decimal, int) and round_decimal > 0:
-            all_scores = [round(x, round_decimal) for x in res.scores]
-        else:
-            all_scores = res.scores
-
-        data = []
-        nq_thres = 0
-        for topk in all_topks:
-            start, end = nq_thres, nq_thres + topk
-            nq_th_fields = self.get_fields_by_range(start, end, fields_data)
-            data.append(
-                Hits(
-                    topk,
-                    all_pks[start:end],
-                    all_scores[start:end],
-                    nq_th_fields,
-                    output_fields,
-                    self._pk_name,
-                )
-            )
-            nq_thres += topk
-        self._session_ts = session_ts
-        self._search_iterator_v2_results = res.search_iterator_v2_results
-        super().__init__(data)
-
-    def get_session_ts(self):
-        return self._session_ts
-
-    def get_search_iterator_v2_results_info(self):
-        return self._search_iterator_v2_results
-
-    def get_fields_by_range(
-        self, start: int, end: int, all_fields_data: List[schema_pb2.FieldData]
-    ) -> Dict[str, Tuple[List[Any], schema_pb2.FieldData]]:
-        field2data: Dict[str, Tuple[List[Any], schema_pb2.FieldData]] = {}
-
-        for field in all_fields_data:
-            name, scalars, dtype = field.field_name, field.scalars, field.type
-            field_meta = schema_pb2.FieldData(
-                type=dtype,
-                field_name=name,
-                field_id=field.field_id,
-                is_dynamic=field.is_dynamic,
-            )
-            if dtype == DataType.BOOL:
-                field2data[name] = (
-                    apply_valid_data(
-                        scalars.bool_data.data[start:end], field.valid_data, start, end
-                    ),
-                    field_meta,
-                )
-                continue
-
-            if dtype in (DataType.INT8, DataType.INT16, DataType.INT32):
-                field2data[name] = (
-                    apply_valid_data(
-                        scalars.int_data.data[start:end], field.valid_data, start, end
-                    ),
-                    field_meta,
-                )
-                continue
-
-            if dtype == DataType.INT64:
-                field2data[name] = (
-                    apply_valid_data(
-                        scalars.long_data.data[start:end], field.valid_data, start, end
-                    ),
-                    field_meta,
-                )
-                continue
-
-            if dtype == DataType.FLOAT:
-                field2data[name] = (
-                    apply_valid_data(
-                        scalars.float_data.data[start:end], field.valid_data, start, end
-                    ),
-                    field_meta,
-                )
-                continue
-
-            if dtype == DataType.DOUBLE:
-                field2data[name] = (
-                    apply_valid_data(
-                        scalars.double_data.data[start:end], field.valid_data, start, end
-                    ),
-                    field_meta,
-                )
-                continue
-
-            if dtype == DataType.VARCHAR:
-                field2data[name] = (
-                    apply_valid_data(
-                        scalars.string_data.data[start:end], field.valid_data, start, end
-                    ),
-                    field_meta,
-                )
-                continue
-
-            if dtype == DataType.JSON:
-                res = apply_valid_data(
-                    scalars.json_data.data[start:end], field.valid_data, start, end
-                )
-                json_dict_list = [ujson.loads(item) if item is not None else item for item in res]
-                field2data[name] = json_dict_list, field_meta
-                continue
-
-            if dtype == DataType.ARRAY:
-                res = apply_valid_data(
-                    scalars.array_data.data[start:end], field.valid_data, start, end
-                )
-                field2data[name] = (
-                    extract_array_row_data(res, scalars.array_data.element_type),
-                    field_meta,
-                )
-                continue
-
-            # vectors
-            dim, vectors = field.vectors.dim, field.vectors
-            field_meta.vectors.dim = dim
-            if dtype == DataType.FLOAT_VECTOR:
-                if start == 0 and (end - start) * dim >= len(vectors.float_vector.data):
-                    # If the range equals to the length of vectors.float_vector.data, direct return
-                    # it to avoid a copy. This logic improves performance by 25% for the case
-                    # retrival 1536 dim embeddings with topk=16384.
-                    field2data[name] = vectors.float_vector.data, field_meta
-                else:
-                    field2data[name] = (
-                        vectors.float_vector.data[start * dim : end * dim],
-                        field_meta,
-                    )
-                continue
-
-            if dtype == DataType.BINARY_VECTOR:
-                field2data[name] = (
-                    vectors.binary_vector[start * (dim // 8) : end * (dim // 8)],
-                    field_meta,
-                )
-                continue
-            # TODO(SPARSE): do we want to allow the user to specify the return format?
-            if dtype == DataType.SPARSE_FLOAT_VECTOR:
-                field2data[name] = (
-                    entity_helper.sparse_proto_to_rows(vectors.sparse_float_vector, start, end),
-                    field_meta,
-                )
-                continue
-
-            if dtype == DataType.BFLOAT16_VECTOR:
-                field2data[name] = (
-                    vectors.bfloat16_vector[start * (dim * 2) : end * (dim * 2)],
-                    field_meta,
-                )
-                continue
-
-            if dtype == DataType.FLOAT16_VECTOR:
-                field2data[name] = (
-                    vectors.float16_vector[start * (dim * 2) : end * (dim * 2)],
-                    field_meta,
-                )
-                continue
-
-            if dtype == DataType.INT8_VECTOR:
-                field2data[name] = (
-                    vectors.int8_vector[start * dim : end * dim],
-                    field_meta,
-                )
-                continue
-        return field2data
-
-    def __str__(self) -> str:
-        """Only print at most 10 query results"""
-        reminder = f" ... and {len(self) - 10} results remaining" if len(self) > 10 else ""
-        recall_msg = (
-            f", recalls: {list(map(str, self.recalls[:10]))}" if len(self.recalls) > 0 else ""
-        ) + (
-            f" ... and {len(self.recalls) - 10} recall results remaining"
-            if len(self.recalls) > 10
-            else ""
-        )
-        cost_msg = f", cost: {self.cost}" if self.cost else ""
-        return f"data: {list(map(str, self[:10]))}{reminder}{recall_msg}{cost_msg}"
-
-    __repr__ = __str__
-
-
-class Hits(list):
-    ids: List[Union[str, int]]
-    distances: List[float]
-
-    def __init__(
-        self,
-        topk: int,
-        pks: List[Union[int, str]],
-        distances: List[float],
-        fields: Dict[str, Tuple[List[Any], schema_pb2.FieldData]],
-        output_fields: List[str],
-        pk_name: str,
-    ):
-        """
-        Args:
-            fields(Dict[str, Tuple[List[Any], schema_pb2.FieldData]]):
-                field name to a tuple of topk data and field meta
-        """
-        self.ids = pks
-        self.distances = distances
-        self._pk_name = pk_name
-
-        all_fields = list(fields.keys())
-        dynamic_fields = list(set(output_fields) - set(all_fields))
-
-        hits = []
-        for i in range(topk):
-            curr_field = {}
-            for fname, (data, field_meta) in fields.items():
-                if len(data) <= i:
-                    curr_field[fname] = None
-                # Get dense vectors
-                if field_meta.type in (
-                    DataType.FLOAT_VECTOR,
-                    DataType.BINARY_VECTOR,
-                    DataType.BFLOAT16_VECTOR,
-                    DataType.FLOAT16_VECTOR,
-                    DataType.INT8_VECTOR,
-                ):
-                    dim = field_meta.vectors.dim
-                    if field_meta.type in [DataType.BINARY_VECTOR]:
-                        dim = dim // 8
-                    elif field_meta.type in [DataType.BFLOAT16_VECTOR, DataType.FLOAT16_VECTOR]:
-                        dim = dim * 2
-                    curr_field[fname] = data[i * dim : (i + 1) * dim]
-                    continue
-
-                # Get dynamic fields
-                if field_meta.type == DataType.JSON and field_meta.is_dynamic:
-                    if len(dynamic_fields) > 0:
-                        curr_field.update({k: v for k, v in data[i].items() if k in dynamic_fields})
-                        continue
-
-                    if fname in output_fields:
-                        curr_field.update(data[i])
-                        continue
-
-                # sparse float vector and other fields
-                curr_field[fname] = data[i]
-
-            hits.append(Hit(pks[i], distances[i], curr_field, self._pk_name))
-
-        super().__init__(hits)
-
-    def __str__(self) -> str:
-        """Only print at most 10 query results"""
-        reminder = f" ... and {len(self) - 10} entities remaining" if len(self) > 10 else ""
-        return f"{list(map(str, self[:10]))!s}{reminder}"
-
-    __repr__ = __str__
-
-
-class Hit:
-    id: Union[int, str]
-    distance: float
-    fields: Dict[str, Any]
-
-    def __init__(self, pk: Union[int, str], distance: float, fields: Dict[str, Any], pk_name: str):
-        self.id = pk
-        self.distance = distance
-        self.fields = fields
-        self._pk_name = pk_name
-
-    def __getattr__(self, item: str):
-        if item not in self.fields:
-            raise MilvusException(message=f"Field {item} is not in the hit entity")
-        return self.fields[item]
-
-    @property
-    def entity(self):
-        return self
-
-    @property
-    def pk(self) -> Union[str, int]:
-        return self.id
-
-    @property
-    def score(self) -> float:
-        return self.distance
-
-    def get(self, field_name: str) -> Any:
-        return self.fields.get(field_name)
-
-    def __str__(self) -> str:
-        return f"{self._pk_name}: {self.id}, distance: {self.distance}, entity: {self.fields}"
-
-    __repr__ = __str__
-
-    def to_dict(self):
-        return {
-            self._pk_name: self.id,
-            "distance": self.distance,
-            "entity": self.fields,
-        }
-
-
-def extract_array_row_data(
-    scalars: List[schema_pb2.ScalarField], element_type: DataType
-) -> List[List[Any]]:
-    row = []
-    for ith_array in scalars:
-        if ith_array is None:
-            row.append(None)
-            continue
-
-        if element_type == DataType.INT64:
-            row.append(ith_array.long_data.data)
-            continue
-
-        if element_type == DataType.BOOL:
-            row.append(ith_array.bool_data.data)
-            continue
-
-        if element_type in (DataType.INT8, DataType.INT16, DataType.INT32):
-            row.append(ith_array.int_data.data)
-            continue
-
-        if element_type == DataType.FLOAT:
-            row.append(ith_array.float_data.data)
-            continue
-
-        if element_type == DataType.DOUBLE:
-            row.append(ith_array.double_data.data)
-            continue
-
-        if element_type in (DataType.STRING, DataType.VARCHAR):
-            row.append(ith_array.string_data.data)
-            continue
-    return row
-
-
-def apply_valid_data(
-    data: List[Any], valid_data: Union[None, List[bool]], start: int, end: int
-) -> List[Any]:
-    if valid_data:
-        for i, valid in enumerate(valid_data[start:end]):
-            if not valid:
-                data[i] = None
-    return data
 
 
 class LoopBase:
