+++ pymilvus/milvus_client/async_milvus_client.py
@@ -6,9 +6,7 @@ from pymilvus.client.abstract import AnnSearchRequest, BaseRanker
 from pymilvus.client.constants import DEFAULT_CONSISTENCY_LEVEL
 from pymilvus.client.types import (
     ExceptionsMessage,
-    ExtraList,
     OmitZeroDict,
-    construct_cost_extra,
 )
 from pymilvus.exceptions import (
     DataTypeNotMatchException,
@@ -360,12 +358,7 @@ class AsyncMilvusClient:
         except Exception as ex:
             logger.error("Failed to hybrid search collection: %s", collection_name)
             raise ex from ex
-
-        ret = []
-        for hits in res:
-            ret.append([hit.to_dict() for hit in hits])
-
-        return ExtraList(ret, extra=construct_cost_extra(res.cost))
+        return res
 
     async def search(
         self,
@@ -398,15 +391,7 @@ class AsyncMilvusClient:
         except Exception as ex:
             logger.error("Failed to search collection: %s", collection_name)
             raise ex from ex
-
-        ret = []
-        for hits in res:
-            query_result = []
-            for hit in hits:
-                query_result.append(hit.to_dict())
-            ret.append(query_result)
-
-        return ExtraList(ret, extra=construct_cost_extra(res.cost), recalls=res.recalls)
+        return res
 
     async def query(
         self,
