+++ pymilvus/orm/collection.py
@@ -17,8 +17,9 @@ from typing import Dict, List, Optional, Union
 import pandas as pd
 
 from pymilvus.client import utils
-from pymilvus.client.abstract import BaseRanker, SearchResult
+from pymilvus.client.abstract import BaseRanker
 from pymilvus.client.constants import DEFAULT_CONSISTENCY_LEVEL
+from pymilvus.client.search_reasult import SearchResult
 from pymilvus.client.types import (
     CompactionPlans,
     CompactionState,
